# Conflicts:
#	src/Nethermind/Nethermind.Consensus.AuRa/AuRaContractGasLimitOverride.cs
#	src/Nethermind/Nethermind.Consensus.Test/ManualGasLimitCalculator.cs
#	src/Nethermind/Nethermind.Consensus.Test/TargetAdjustedGasLimitCalculatorTests.cs
#	src/Nethermind/Nethermind.Consensus/FollowOtherMiners.cs
#	src/Nethermind/Nethermind.Consensus/IGasLimitCalculator.cs
#	src/Nethermind/Nethermind.Consensus/Producers/PayloadAttributes.cs
#	src/Nethermind/Nethermind.Consensus/TargetAdjustedGasLimitCalculator.cs
#	src/Nethermind/Nethermind.Merge.Plugin/BlockProduction/Boost/BoostPayloadAttributes.cs
#	src/Nethermind/Nethermind.Optimism/OptimismGasLimitCalculator.cs
#	src/Nethermind/Nethermind.Optimism/Rpc/OptimismPayloadAttributes.cs
#	src/Nethermind/Nethermind.Taiko/TaikoPayloadAttributes.cs
#	src/Nethermind/Nethermind.Xdc/XdcGasLimitCalculator.cs

+++ tools/Evm/T8n/JsonTypes/Ommer.cs
@@ -5,4 +5,4 @@ using Nethermind.Core;
 
 namespace Evm.T8n.JsonTypes;
 
-public readonly record struct Ommer(int Delta, Address Address);
+public readonly record struct Ommer(ulong Delta, Address Address);
